+++ packages/react/src/constants/index.ts
@@ -197,3 +197,8 @@ export const NEAR_NETWORK_CONFIG: Record<string, string> = {
   testnet: CHAIN_ID.nearTestnet,
   mainnet: CHAIN_ID.near,
 };
+
+/**
+ * The native SLF token.
+ */
+export const SLF_TOKEN = 'uslf';