+++ packages/react/src/hooks/useBitcoinStore.ts
@@ -5,6 +5,6 @@ import { BitcoinState } from '../store/Bitcoin.js';
 
 export function useBitcoinStore<T>(selector: (state: BitcoinState) => T): T {
   const { store } = useContext(BitcoinContext);
-  if (!store) throw new Error('Missing Ton Provider in the tree');
+  if (!store) throw new Error('Missing Bitcoin Provider in the tree');
   return useStore(store, selector);
 }